Combined two readCSV modules into one
